feat : convert odds and money to int

+++ lib_database/src/main/java/arch/cayenne/lib/database/entity/BetBean.kt
@@ -9,7 +9,7 @@ data class BetBean(
     val gameId: Int,
     val betTeamName: String,
     var handicap: String,
-    var odds: Float,
+    var odds: Int,
     var betType: BetTypeEnum, // 0: 單注 1: 串關 2: 預約
     val leagueName: String,
     val matchName: String,
